# public/denis23x.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "description": "This API provides access to a collection of books, users, and reviews, allowing clients to perform various operations like viewing available books, creating and managing users, and posting or retrieving book reviews.\n\n- **Books**: Retrieve book details, search for books, and view books.\n- **Users**: Manage users, view user profiles.\n- **Reviews**: Add, view, update, or delete reviews for books.",
    "title": "Books Showcase Demo",
    "contact": {
      "email": "denix23x@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/api/v1/placeholder/users": {
      "get": {
        "summary": "Get All",
        "deprecated": false,
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "number",
              "default": 1,
              "examples": [
                1
              ],
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "example": 10,
            "schema": {
              "type": "number",
              "default": 10,
              "examples": [
                10
              ],
              "minimum": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": [
                        {
                          "id": 827,
                          "avatar": "https://picsum.photos/seed/rimvSVV/512/512?blur=1",
                          "firstName": "Shea",
                          "lastName": "Haag",
                          "gender": "Trans woman",
                          "sex": "Male",
                          "bio": "case devotee  💂🏻‍♂️",
                          "email": "Claud.Pfeffer77@yahoo.com",
                          "createdAt": "2024-01-03T00:47:24.715Z",
                          "updatedAt": "2024-11-13T18:58:01.802Z"
                        },
                        {
                          "id": 325,
                          "avatar": "https://picsum.photos/seed/oiFrh/512/512?blur=1",
                          "firstName": "Reginald",
                          "lastName": "Leffler-Romaguera",
                          "gender": "Genderflux",
                          "sex": "Male",
                          "bio": "student",
                          "email": "Dandre_Zemlak@hotmail.com",
                          "createdAt": "2024-08-18T04:33:44.062Z",
                          "updatedAt": "2024-11-14T09:11:16.430Z"
                        },
                        {
                          "id": 164,
                          "avatar": "https://picsum.photos/seed/oJtaJ/512/512?blur=1",
                          "firstName": "Marc",
                          "lastName": "Leannon",
                          "gender": "Two* person",
                          "sex": "Male",
                          "bio": "streamer, designer, traveler",
                          "email": "Earl43@yahoo.com",
                          "createdAt": "2024-08-02T08:31:58.707Z",
                          "updatedAt": "2024-11-14T07:23:46.363Z"
                        }
                      ],
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "Create User",
        "deprecated": false,
        "description": "",
        "tags": [
          "Users"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "description": "User Avatar",
                    "nullable": true
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "description": "User Gender"
                  },
                  "sex": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female"
                    ],
                    "examples": [
                      "Male"
                    ]
                  },
                  "bio": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "avatar": "https://picsum.photos/seed/xw4U3l/512/512",
                "firstName": "Ethel",
                "lastName": "Bauch",
                "gender": "Bigender",
                "sex": "female",
                "bio": "engineer, teacher",
                "email": "Ethel_Bauch@hotmail.com"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 960,
                        "avatar": "https://picsum.photos/seed/wCDhmwW/512/512?blur=1",
                        "firstName": "Lera",
                        "lastName": "Konopelski-Jaskolski",
                        "gender": "Trans female",
                        "sex": "Male",
                        "bio": "public speaker",
                        "email": "Candice.Runolfsson@yahoo.com",
                        "createdAt": "2023-12-19T03:40:32.324Z",
                        "updatedAt": "2024-11-14T07:15:58.497Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "issues": [
                          {
                            "code": "too_small",
                            "minimum": 2,
                            "type": "string",
                            "inclusive": true,
                            "exact": false,
                            "message": "String must contain at least 2 character(s)",
                            "path": [
                              "title"
                            ]
                          }
                        ],
                        "name": "ZodError"
                      },
                      "status": 400
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id",
            "minimum": 1
          },
          "avatar": {
            "type": "string",
            "description": "User Avatar",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "description": "User Gender"
          },
          "sex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "examples": [
              "Male"
            ]
          },
          "bio": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": false
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Develop Env"
    },
    {
      "url": "https://test.your-api-server.com",
      "description": "Testing Env"
    }
  ],
...
}